+++ packages/framework/src/Model/Customer/UserDataFactory.php
@@ -72,5 +72,6 @@ class UserDataFactory implements UserDataFactoryInterface
         $userData->email = $user->getEmail();
         $userData->pricingGroup = $user->getPricingGroup();
         $userData->createdAt = $user->getCreatedAt();
+        $userData->telephone = $user->getTelephone();
     }
 }